refactor: move user-related models to command package and update imports
