+++ identification/pkg/events/user_create.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/L4B0MB4/EVTSRC/pkg/models"
-	m "github.com/L4B0MB4/PRYVT/identification/pkg/models"
+	m "github.com/L4B0MB4/PRYVT/identification/pkg/models/command"
 )
 
 type UserCreatedEvent struct {
