+++ identification/pkg/models/command/change_name.go
@@ -1,4 +1,4 @@
-package models
+package command
 
 type ChangeDisplayName struct {
 	DisplayName string
