+++ identification/pkg/models/command/user_id.go
@@ -1,4 +1,4 @@
-package models
+package command
 
 type UserIdInfo struct {
 	UserId string
