+++ identification/pkg/models/query/user_info.go
@@ -1,4 +1,4 @@
-package models
+package query
 
 import (
 	"time"
